# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Event API",
    "version": "1.0.0",
    "description": "This API allows users to manage events and address events.",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/routesEvents": {
      "get": {
        "summary": "Get all routes events",
        "operationId": "getAllRoutesEvents",
        "description": "This endpoint retrieves a list of all routes events stored in the system. Routes events represent specific route points related to an event, and are crucial for mapping and planning purposes. If no routes events are found, a 404 error will be returned.",
        "tags": [
          "RoutesEvents"
        ],
        "responses": {
          "200": {
            "description": "A list of routes events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoutesEvent"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No routes events found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a new routes event",
        "operationId": "createRoutesEvent",
        "description": "This endpoint allows you to create a new routes event by providing the required parameters such as latitude, longitude, order, and the related address event ID. If any of these fields are missing, a 400 error will be returned.",
        "tags": [
          "RoutesEvents"
        ],
        "requestBody": {
          "description": "Routes event to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutesEventInput"
              },
              "example": {
                "latitude": 37.7749,
                "longitude": -122.4194,
                "order": 1,
                "addressEvent_id": "dddddddd-dddd-dddd-dddd-dddddddddddd"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Routes event created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutesEvent"
                },
                "example": {
                  "routeID": "f8d99d7e-5cfd-42d7-9fdd-c33f33c9a413",
                  "latitude": 37.7749,
                  "longitude": -122.4194,
                  "order": 1,
                  "addressEvent_id": "dddddddd-dddd-dddd-dddd-dddddddddddd"
                }
              }
            },
            "400": {
              "description": "Bad Request - Missing required fields"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      }
    },
    "/routesEvents/{id}": {
      "get": {
        "summary": "Get a specific routes event by ID",
        "operationId": "getRoutesEventById",
        "description": "This endpoint allows you to retrieve a specific routes event using its unique ID. If the routes event does not exist, a 404 error will be returned.",
        "tags": [
          "RoutesEvents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the routes event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Routes event found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutesEvent"
                }
              }
            }
          },
          "404": {
            "description": "Routes event not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "summary": "Update a specific routes event by ID",
        "operationId": "updateRoutesEvent",
        "description": "This endpoint allows you to update the details of a specific routes event, such as latitude, longitude, order, and the related address event ID. If the routes event is not found, a 404 error will be returned.",
        "tags": [
          "RoutesEvents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the routes event",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Updated routes event data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoutesEventInput"
              },
              "example": {
                "latitude": 37.7749,
                "longitude": -122.4194,
                "order": 2,
                "addressEvent_id": "dddddddd-dddd-dddd-dddd-dddddddddddd"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Routes event updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoutesEvent"
                }
              }
            }
          },
          "404": {
            "description": "Routes event not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific routes event by ID",
        "operationId": "deleteRoutesEvent",
        "description": "This endpoint allows you to delete a specific routes event by ID. If the routes event is not found, a 404 error will be returned. If the deletion is successful, it returns a 204 status code.",
        "tags": [
          "RoutesEvents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the routes event to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Routes event deleted successfully"
          },
          "404": {
            "description": "Routes event not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Bearer token authentication"
      }
    },
    "schemas": {
...
      "RoutesEvent": {
        "type": "object",
        "properties": {
          "routeID": {
            "type": "string",
            "description": "ID of the route event"
          },
          "latitude": {
            "type": "number",
            "description": "Latitude of the route"
          },
          "longitude": {
            "type": "number",
            "description": "Longitude of the route"
          },
          "order": {
            "type": "integer",
            "description": "Order of the route in the sequence"
          },
          "addressEvent_id": {
            "type": "string",
            "description": "ID of the associated address event"
          }
        }
      },
      "RoutesEventInput": {
        "type": "object",
        "required": [
          "latitude",
          "longitude",
          "order",
          "addressEvent_id"
        ],
        "properties": {
          "latitude": {
            "type": "number",
            "description": "Latitude of the route"
          },
          "longitude": {
            "type": "number",
            "description": "Longitude of the route"
          },
          "order": {
            "type": "integer",
            "description": "Order of the route in the sequence"
          },
          "addressEvent_id": {
            "type": "string",
            "description": "ID of the associated address event"
          }
        }
      },
...
    }
  }
}